landing/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
google-api-core==2.28.1
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.25.2
idna==3.11
inflection==0.5.1
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.2.1
oauthlib==3.3.1
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyairtable==2.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.0
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
